Files renamed
see #1
