+++ gulp/gulpFile.js
@@ -2,9 +2,7 @@ require('./watch.js');
 require('./build.js');
 require('./clean.js');
 require('./jshint.js');
-
-require('./meta-option');
-
+require('./metaOption');
 require('./handlebars.js');
 require('./copy.js');
 require('./less.js');
